Refactor: Make MemoryCache generic
